+++ tests/test_import.py
@@ -9,7 +9,7 @@ from ubelt.util_import import PythonPathContext
 
 def test_import_modpath_basic():
     assert 'testmod' not in sys.modules
-    with pytest.warns(DeprecationWarning):
+    with pytest.warns(Warning):
         temp = ub.TempDir()
     with temp:
         modpath = ub.Path(temp.dpath) / 'testmod.py'
@@ -29,7 +29,7 @@ def test_import_modpath_basic():
 
 def test_import_modpath_package():
     assert '_tmproot373.sub1.sub2.testmod' not in sys.modules
-    with pytest.warns(DeprecationWarning):
+    with pytest.warns(Warning):
         temp = ub.TempDir().start()
     # with ub.TempDir() as temp:
     if True:
@@ -81,7 +81,7 @@ def _static_modname_to_modpath(modname, **kwargs):
 
 
 def test_modname_to_modpath_single():
-    with pytest.warns(DeprecationWarning):
+    with pytest.warns(Warning):
         temp = ub.TempDir()
     with temp:
         dpath = temp.dpath
@@ -108,7 +108,7 @@ def test_modname_to_modpath_package():
     CommandLine:
         pytest testing/test_static.py::test_modname_to_modpath_package
     """
-    with pytest.warns(DeprecationWarning):
+    with pytest.warns(Warning):
         temp = ub.TempDir()
     with temp:
         dpath = temp.dpath
@@ -237,7 +237,7 @@ def test_modname_to_modpath_namespace():
     %timeit _syspath_modname_to_modpath('xdoctest.static_analysis')
     %timeit _importlib_modname_to_modpath('xdoctest.static_analysis')
     """
-    with pytest.warns(DeprecationWarning):
+    with pytest.warns(Warning):
         temp = ub.TempDir()
     with temp:
         dpath = temp.dpath
@@ -276,7 +276,7 @@ def test_package_submodules():
         xdoctest -m ~/code/ubelt/tests/test_import.py test_package_submodules
     """
     from xdoctest import static_analysis as static
-    with pytest.warns(DeprecationWarning):
+    with pytest.warns(Warning):
         temp = ub.TempDir()
     with temp:
         dpath = temp.dpath
@@ -338,7 +338,7 @@ def test_modpath_to_modname():
         pytest testing/test_static.py::test_modpath_to_modname -s
         python testing/test_static.py test_modpath_to_modname
     """
-    with pytest.warns(DeprecationWarning):
+    with pytest.warns(Warning):
         temp = ub.TempDir()
     with temp:
         dpath = temp.dpath